--- pyEDAA/OutputFilter/Xilinx/Exception.py
@@ -1,78 +1,0 @@
-# ==================================================================================================================== #
-#               _____ ____    _        _      ___        _               _   _____ _ _ _                               #
-#   _ __  _   _| ____|  _ \  / \      / \    / _ \ _   _| |_ _ __  _   _| |_|  ___(_) | |_ ___ _ __                    #
-#  | '_ \| | | |  _| | | | |/ _ \    / _ \  | | | | | | | __| '_ \| | | | __| |_  | | | __/ _ \ '__|                   #
-#  | |_) | |_| | |___| |_| / ___ \  / ___ \ | |_| | |_| | |_| |_) | |_| | |_|  _| | | | ||  __/ |                      #
-#  | .__/ \__, |_____|____/_/   \_\/_/   \_(_)___/ \__,_|\__| .__/ \__,_|\__|_|   |_|_|\__\___|_|                      #
-#  |_|    |___/                                             |_|                                                        #
-# ==================================================================================================================== #
-# Authors:                                                                                                             #
-#   Patrick Lehmann                                                                                                    #
-#                                                                                                                      #
-# License:                                                                                                             #
-# ==================================================================================================================== #
-# Copyright 2025-2026 Electronic Design Automation Abstraction (EDA²)                                                  #
-#                                                                                                                      #
-# Licensed under the Apache License, Version 2.0 (the "License");                                                      #
-# you may not use this file except in compliance with the License.                                                     #
-# You may obtain a copy of the License at                                                                              #
-#                                                                                                                      #
-#   http://www.apache.org/licenses/LICENSE-2.0                                                                         #
-#                                                                                                                      #
-# Unless required by applicable law or agreed to in writing, software                                                  #
-# distributed under the License is distributed on an "AS IS" BASIS,                                                    #
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.                                             #
-# See the License for the specific language governing permissions and                                                  #
-# limitations under the License.                                                                                       #
-#                                                                                                                      #
-# SPDX-License-Identifier: Apache-2.0                                                                                  #
-# ==================================================================================================================== #
-#
-"""Basic classes for outputs from AMD/Xilinx Vivado."""
-from pyTooling.Decorators import export
-
-from pyEDAA.OutputFilter  import OutputFilterException
-
-
-@export
-class ProcessorException(OutputFilterException):
-	"""
-	Base-class for exceptions raised by processors parsing log outputs.
-	"""
-
-
-@export
-class ClassificationException(ProcessorException):
-	"""
-	Raised if a log output line couldn't be classified.
-	"""
-	_lineNumber: int  #: Line number of the unclassified line.
-	_rawMessage: str  #: Raw message of the unclassified line.
-
-	def __init__(self, errorMessage: str, lineNumber: int, rawMessageLine: str) -> None:
-		"""
-		Initializes a classification exception.
-
-		:param errorMessage:   Error message why the line couldn't be classified.
-		:param lineNumber:     Line number of the unclassified line.
-		:param rawMessageLine: Raw message of the unclassified line.
-		"""
-		super().__init__(errorMessage)
-
-		self._lineNumber = lineNumber
-		self._rawMessage = rawMessageLine
-
-	def __str__(self) -> str:
-		return f"{self.message}: {self._rawMessage} (at line {self._lineNumber})"
-
-
-@export
-class ParserStateException(ProcessorException):
-	"""
-	Raised if a log output parser has a broken state.
-	"""
-
-
-@export
-class NotPresentException(ProcessorException):
-	pass
