+++ pyEDAA/OutputFilter/Xilinx/SynthesizeDesign.py
@@ -30,250 +30,17 @@
 #
 """A filtering anc classification processor for AMD/Xilinx Vivado Synthesis outputs."""
 from re     import compile as re_compile
-from typing import ClassVar, Dict, Generator, Type, Optional as Nullable, Any, Union
+from typing import ClassVar, Dict, Generator, Optional as Nullable
 
-from pyTooling.Common      import getFullyQualifiedName
 from pyTooling.Decorators  import export, readonly
-from pyTooling.MetaClasses import ExtendedType, abstractmethod
 
-from pyEDAA.OutputFilter.Xilinx.Exception import NotPresentException
-from pyEDAA.OutputFilter.Xilinx.Common    import VHDLAssertionMessage, Line, LineKind, VivadoInfoMessage, VHDLReportMessage, VivadoMessage
-from pyEDAA.OutputFilter.Xilinx.Common2   import BaseParser
+from pyEDAA.OutputFilter.Xilinx import Section, SubSection, SectionWithChildren
+from pyEDAA.OutputFilter.Xilinx import VHDLAssertionMessage, VivadoLine, LineKind, VivadoInfoMessage, VHDLReportMessage, VivadoMessage
+
+
+__all__ = ["TIME_MEMORY_PATTERN"]
 
 TIME_MEMORY_PATTERN = re_compile(r"""Time \(s\): cpu = (\d{2}:\d{2}:\d{2}) ; elapsed = (\d{2}:\d{2}:\d{2}) . Memory \(MB\): peak = (\d+\.\d+) ; gain = (\d+\.\d+)""")
-
-
-@export
-class SubSectionNotPresentException(NotPresentException):
-	pass
-
-
-@export
-class BaseSection(metaclass=ExtendedType, mixin=True):
-	@abstractmethod
-	def _SectionStart(self, line: Line) -> Generator[Line, Line, Line]:
-		pass
-
-	@abstractmethod
-	def _SectionFinish(self, line: Line) -> Generator[Line, Line, None]:
-		pass
-
-	@abstractmethod
-	def Generator(self, line: Line) -> Generator[Line, Line, Line]:
-		pass
-
-
-@export
-class Section(BaseParser, BaseSection):
-	"""
-	Base-class for sections within log outputs from *synthesize design*.
-	"""
-	# _NAME:   ClassVar[str]
-	# _START:  ClassVar[str]
-	# _FINISH: ClassVar[str]
-
-	_command:  "Command"  #: Reference to the command (parent).
-	_duration: float      #: Duration synthesis spent in processing a synthesis step logged in this log output section.
-
-	def __init__(self, command: "Command") -> None:
-		"""
-		Initialized a section.
-
-		:param command: Reference to the parent TCL command.
-		"""
-		super().__init__()  #command._processor)
-
-		self._command = command
-		self._duration = 0.0
-
-	@readonly
-	def Duration(self) -> float:
-		"""
-		Read-only property to access the duration synthesis spent in processing a synthesis step logged in this log output
-		section.
-
-		:returns: Synthesis step duration in seconds.
-		"""
-		return self._duration
-
-	def _SectionStart(self, line: Line) -> Generator[Line, Line, Line]:
-		line._kind = LineKind.SectionStart
-
-		line = yield line
-		if line.StartsWith("----"):
-			line._kind = LineKind.SectionStart | LineKind.SectionDelimiter
-		else:
-			line._kind |= LineKind.ProcessorError
-
-		nextLine = yield line
-		return nextLine
-
-	def _SectionFinish(self, line: Line, skipDashes: bool = False) -> Generator[Line, Line, None]:
-		if not skipDashes:
-			if line.StartsWith("----"):
-				line._kind = LineKind.SectionEnd | LineKind.SectionDelimiter
-			else:
-				line._kind |= LineKind.ProcessorError
-
-			line = yield line
-
-		if line.StartsWith(self._FINISH):
-			line._kind = LineKind.SectionEnd
-		else:
-			line._kind |= LineKind.ProcessorError
-
-		line = yield line
-		if line.StartsWith("----"):
-			line._kind = LineKind.SectionEnd | LineKind.SectionDelimiter
-		else:
-			line._kind |= LineKind.ProcessorError
-
-		nextLine = yield line
-		return nextLine
-
-	# @mustoverride
-	# def ParseLine(self, lineNumber: int, line: str) -> ProcessingState:
-	# 	if len(line) == 0:
-	# 		return ProcessingState.EmptyLine
-	# 	elif line.startswith("----"):
-	# 		return ProcessingState.DelimiterLine
-	# 	elif line.startswith(self._START):
-	# 		return ProcessingState.Skipped
-	# 	elif line.startswith(self._FINISH):
-	# 		l = line[len(self._FINISH):]
-	# 		if (match := TIME_MEMORY_PATTERN.match(l)) is not None:
-	# 			# cpuParts = match[1].split(":")
-	# 			elapsedParts = match[2].split(":")
-	# 			# peakMemory = float(match[3])
-	# 			# gainMemory = float(match[4])
-	# 			self._duration = int(elapsedParts[0]) * 3600 + int(elapsedParts[1]) * 60 + int(elapsedParts[2])
-	#
-	# 		return ProcessingState.Skipped | ProcessingState.Last
-	# 	elif line.startswith("Start") or line.startswith("Starting"):
-	# 		print(f"ERROR: didn't find finish\n  {line}")
-	# 		return ProcessingState.Reprocess
-	#
-	# 	return ProcessingState.Skipped
-
-	def Generator(self, line: Line) -> Generator[Line, Line, Line]:
-		line = yield from self._SectionStart(line)
-
-		while True:
-			if line._kind is LineKind.Empty:
-				line = yield line
-				continue
-			elif line.StartsWith("----"):
-				line._kind = LineKind.SectionEnd | LineKind.SectionDelimiter
-				break
-			elif isinstance(line, VivadoMessage):
-				self._AddMessage(line)
-			else:
-				line._kind = LineKind.Verbose
-
-			line = yield line
-
-		# line = yield line
-		nextLine = yield from self._SectionFinish(line)
-		return nextLine
-
-
-@export
-class SubSection(BaseParser, BaseSection):
-	"""
-	Base-class for subsections within log outputs from *synthesize design*.
-	"""
-	# _NAME:    ClassVar[str]
-
-	_section: Section  #: Reference to the section (parent).
-
-	def __init__(self, section: Section) -> None:
-		"""
-		Initialized a subsection.
-
-		:param section: Reference to the parent section.
-		"""
-		super().__init__()
-		self._section = section
-
-	def _SectionStart(self, line: Line) -> Generator[Line, Line, Line]:
-		line._kind = LineKind.SubSectionStart
-
-		line = yield line
-		if line.StartsWith("----"):
-			line._kind = LineKind.SubSectionStart | LineKind.SubSectionDelimiter
-		else:
-			line._kind |= LineKind.ProcessorError
-
-		nextLine = yield line
-		return nextLine
-
-	def _SectionFinish(self, line: Line) -> Generator[Line, Line, None]:
-		if line.StartsWith("----"):
-			line._kind = LineKind.SubSectionEnd | LineKind.SubSectionDelimiter
-		else:
-			line._kind |= LineKind.ProcessorError
-
-		line = yield line
-		if line.StartsWith(self._FINISH):
-			line._kind = LineKind.SubSectionEnd
-		else:
-			line._kind |= LineKind.ProcessorError
-
-		line = yield line
-		if line.StartsWith("----"):
-			line._kind = LineKind.SubSectionEnd | LineKind.SubSectionDelimiter
-		else:
-			line._kind |= LineKind.ProcessorError
-
-		nextLine = yield line
-		return nextLine
-
-	def Generator(self, line: Line) -> Generator[Line, Line, Line]:
-		line = yield from self._SectionStart(line)
-
-		while True:
-			if line._kind is LineKind.Empty:
-				line = yield line
-				continue
-			elif line.StartsWith("----"):
-				line._kind = LineKind.SubSectionEnd | LineKind.SubSectionDelimiter
-				break
-			elif isinstance(line, VivadoMessage):
-				self._AddMessage(line)
-			else:
-				line._kind = LineKind.Verbose
-
-			line = yield line
-
-		nextLine = yield from self._SectionFinish(line)
-		return nextLine
-
-
-@export
-class SectionWithChildren(Section):
-	"""
-	Base-class for sections with subsections.
-	"""
-	_subsections: Dict[Type[SubSection], SubSection]
-
-	def __init__(self, command: "Command") -> None:
-		super().__init__(command)
-
-		self._subsections = {}
-
-	def __contains__(self, key: Any) -> bool:
-		if not issubclass(key, SubSection):
-			ex = TypeError(f"Parameter 'item' is not a SubSection.")
-			ex.add_note(f"Got type '{getFullyQualifiedName(key)}'.")
-			raise ex
-
-		return key in self._subsections
-
-	def __getitem__(self, item: Type[SubSection]) -> SubSection:
-		try:
-			return self._sections[item]
-		except KeyError as ex:
-			raise SubSectionNotPresentException(f"SubSection '{item._NAME}' not present in '{self._parent._parent.logfile}'.") from ex
 
 
 @export
@@ -283,11 +50,12 @@ class RTLElaboration(Section):
 
 	Used by Vivado command :class:`~pyEDAA.OutputFilter.Xilinx.Commands.SynthesizeDesign`.
 	"""
-	_NAME:   ClassVar[str] = "RTL Elaboration"
-	_START:  ClassVar[str] = "Starting RTL Elaboration : "
-	_FINISH: ClassVar[str] = "Finished RTL Elaboration : "
+	_NAME:       ClassVar[str] = "RTL Elaboration"
+	_START:      ClassVar[str] = "Starting RTL Elaboration : "
+	_FINISH:     ClassVar[str] = "Finished RTL Elaboration : "
+	_DUPLICATES: ClassVar[bool] = False
 
-	def Generator(self, line: Line) -> Generator[Line, Line, Line]:
+	def Generator(self, line: VivadoLine) -> Generator[VivadoLine, VivadoLine, VivadoLine]:
 		line = yield from self._SectionStart(line)
 
 		while True:
@@ -331,9 +99,10 @@ class HandlingCustomAttributes(Section):
 
 	Used by Vivado command :class:`~pyEDAA.OutputFilter.Xilinx.Commands.SynthesizeDesign`.
 	"""
-	_NAME:   ClassVar[str] = "Handling Custom Attributes"
-	_START:  ClassVar[str] = "Start Handling Custom Attributes"
-	_FINISH: ClassVar[str] = "Finished Handling Custom Attributes : "
+	_NAME:       ClassVar[str] = "Handling Custom Attributes"
+	_START:      ClassVar[str] = "Start Handling Custom Attributes"
+	_FINISH:     ClassVar[str] = "Finished Handling Custom Attributes : "
+	_DUPLICATES: ClassVar[bool] = True
 
 
 @export
@@ -343,9 +112,10 @@ class ConstraintValidation(Section):
 
 	Used by Vivado command :class:`~pyEDAA.OutputFilter.Xilinx.Commands.SynthesizeDesign`.
 	"""
-	_NAME:   ClassVar[str] = "Constraint Validation"
-	_START:  ClassVar[str] = "Finished RTL Optimization Phase 1"
-	_FINISH: ClassVar[str] = "Finished Constraint Validation : "
+	_NAME:       ClassVar[str] = "Constraint Validation"
+	_START:      ClassVar[str] = "Finished RTL Optimization Phase 1"
+	_FINISH:     ClassVar[str] = "Finished Constraint Validation : "
+	_DUPLICATES: ClassVar[bool] = False
 
 
 @export
@@ -355,9 +125,10 @@ class LoadingPart(Section):
 
 	Used by Vivado command :class:`~pyEDAA.OutputFilter.Xilinx.Commands.SynthesizeDesign`.
 	"""
-	_NAME:   ClassVar[str] = "Loading Part and Timing Information"
-	_START:  ClassVar[str] = "Start Loading Part and Timing Information"
-	_FINISH: ClassVar[str] = "Finished Loading Part and Timing Information : "
+	_NAME:       ClassVar[str] = "Loading Part and Timing Information"
+	_START:      ClassVar[str] = "Start Loading Part and Timing Information"
+	_FINISH:     ClassVar[str] = "Finished Loading Part and Timing Information : "
+	_DUPLICATES: ClassVar[bool] = False
 
 	_part:   Nullable[str]  #: Part name of the device this design was synthesized for.
 
@@ -380,7 +151,7 @@ class LoadingPart(Section):
 		"""
 		return self._part
 
-	def Generator(self, line: Line) -> Generator[Line, Line, Line]:
+	def Generator(self, line: VivadoLine) -> Generator[VivadoLine, VivadoLine, VivadoLine]:
 		line = yield from self._SectionStart(line)
 
 		while True:
@@ -403,16 +174,18 @@ class LoadingPart(Section):
 		nextLine = yield from self._SectionFinish(line)
 		return nextLine
 
+
 @export
-class ApplySetProperty(Section):
+class ApplySetPropertyXDCConstraints(Section):
 	"""
 	*Applying 'set_property' XDC Constraints* section.
 
 	Used by Vivado command :class:`~pyEDAA.OutputFilter.Xilinx.Commands.SynthesizeDesign`.
 	"""
-	_NAME:   ClassVar[str] = "Applying 'set_property' XDC Constraints"
-	_START:  ClassVar[str] = "Start Applying 'set_property' XDC Constraints"
-	_FINISH: ClassVar[str] = "Finished applying 'set_property' XDC Constraints : "
+	_NAME:       ClassVar[str] = "Applying 'set_property' XDC Constraints"
+	_START:      ClassVar[str] = "Start Applying 'set_property' XDC Constraints"
+	_FINISH:     ClassVar[str] = "Finished applying 'set_property' XDC Constraints : "
+	_DUPLICATES: ClassVar[bool] = False
 
 
 @export
@@ -422,11 +195,12 @@ class RTLComponentStatistics(Section):
 
 	Used by Vivado command :class:`~pyEDAA.OutputFilter.Xilinx.Commands.SynthesizeDesign`.
 	"""
-	_NAME:   ClassVar[str] = "RTL Component Statistics"
-	_START:  ClassVar[str] = "Start RTL Component Statistics"
-	_FINISH: ClassVar[str] = "Finished RTL Component Statistics"
+	_NAME:       ClassVar[str] = "RTL Component Statistics"
+	_START:      ClassVar[str] = "Start RTL Component Statistics"
+	_FINISH:     ClassVar[str] = "Finished RTL Component Statistics"
+	_DUPLICATES: ClassVar[bool] = False
 
-	def Generator(self, line: Line) -> Generator[Line, Line, Line]:
+	def Generator(self, line: VivadoLine) -> Generator[VivadoLine, VivadoLine, VivadoLine]:
 		line = yield from self._SectionStart(line)
 
 		while True:
@@ -454,9 +228,10 @@ class RTLHierarchicalComponentStatistics(Section):
 
 	Used by Vivado command :class:`~pyEDAA.OutputFilter.Xilinx.Commands.SynthesizeDesign`.
 	"""
-	_NAME:   ClassVar[str] = "RTL Hierarchical Component Statistics"
-	_START:  ClassVar[str] = "Start RTL Hierarchical Component Statistics"
-	_FINISH: ClassVar[str] = "Finished RTL Hierarchical Component Statistics"
+	_NAME:       ClassVar[str] = "RTL Hierarchical Component Statistics"
+	_START:      ClassVar[str] = "Start RTL Hierarchical Component Statistics"
+	_FINISH:     ClassVar[str] = "Finished RTL Hierarchical Component Statistics"
+	_DUPLICATES: ClassVar[bool] = False
 
 
 @export
@@ -466,9 +241,10 @@ class PartResourceSummary(Section):
 
 	Used by Vivado command :class:`~pyEDAA.OutputFilter.Xilinx.Commands.SynthesizeDesign`.
 	"""
-	_NAME:   ClassVar[str] = "Part Resource Summary"
-	_START:  ClassVar[str] = "Start Part Resource Summary"
-	_FINISH: ClassVar[str] = "Finished Part Resource Summary"
+	_NAME:       ClassVar[str] = "Part Resource Summary"
+	_START:      ClassVar[str] = "Start Part Resource Summary"
+	_FINISH:     ClassVar[str] = "Finished Part Resource Summary"
+	_DUPLICATES: ClassVar[bool] = False
 
 
 @export
@@ -478,9 +254,10 @@ class CrossBoundaryAndAreaOptimization(Section):
 
 	Used by Vivado command :class:`~pyEDAA.OutputFilter.Xilinx.Commands.SynthesizeDesign`.
 	"""
-	_NAME:   ClassVar[str] = "Cross Boundary and Area Optimization"
-	_START:  ClassVar[str] = "Start Cross Boundary and Area Optimization"
-	_FINISH: ClassVar[str] = "Finished Cross Boundary and Area Optimization : "
+	_NAME:       ClassVar[str] = "Cross Boundary and Area Optimization"
+	_START:      ClassVar[str] = "Start Cross Boundary and Area Optimization"
+	_FINISH:     ClassVar[str] = "Finished Cross Boundary and Area Optimization : "
+	_DUPLICATES: ClassVar[bool] = False
 
 
 @export
@@ -490,9 +267,10 @@ class ROM_RAM_DSP_SR_Retiming(Section):
 
 	Used by Vivado command :class:`~pyEDAA.OutputFilter.Xilinx.Commands.SynthesizeDesign`.
 	"""
-	_NAME:   ClassVar[str] = "ROM, RAM, DSP, Shift Register and Retiming Reporting"
-	_START:  ClassVar[str] = "Start ROM, RAM, DSP, Shift Register and Retiming Reporting"
-	_FINISH: ClassVar[str] = "Finished ROM, RAM, DSP, Shift Register and Retiming Reporting"
+	_NAME:       ClassVar[str] = "ROM, RAM, DSP, Shift Register and Retiming Reporting"
+	_START:      ClassVar[str] = "Start ROM, RAM, DSP, Shift Register and Retiming Reporting"
+	_FINISH:     ClassVar[str] = "Finished ROM, RAM, DSP, Shift Register and Retiming Reporting"
+	_DUPLICATES: ClassVar[bool] = True
 
 
 @export
@@ -502,9 +280,10 @@ class ApplyingXDCTimingConstraints(Section):
 
 	Used by Vivado command :class:`~pyEDAA.OutputFilter.Xilinx.Commands.SynthesizeDesign`.
 	"""
-	_NAME:   ClassVar[str] = "Applying XDC Timing Constraints"
-	_START:  ClassVar[str] = "Start Applying XDC Timing Constraints"
-	_FINISH: ClassVar[str] = "Finished Applying XDC Timing Constraints : "
+	_NAME:       ClassVar[str] = "Applying XDC Timing Constraints"
+	_START:      ClassVar[str] = "Start Applying XDC Timing Constraints"
+	_FINISH:     ClassVar[str] = "Finished Applying XDC Timing Constraints : "
+	_DUPLICATES: ClassVar[bool] = False
 
 
 @export
@@ -514,9 +293,10 @@ class TimingOptimization(Section):
 
 	Used by Vivado command :class:`~pyEDAA.OutputFilter.Xilinx.Commands.SynthesizeDesign`.
 	"""
-	_NAME:   ClassVar[str] = "Timing Optimization"
-	_START:  ClassVar[str] = "Start Timing Optimization"
-	_FINISH: ClassVar[str] = "Finished Timing Optimization : "
+	_NAME:       ClassVar[str] = "Timing Optimization"
+	_START:      ClassVar[str] = "Start Timing Optimization"
+	_FINISH:     ClassVar[str] = "Finished Timing Optimization : "
+	_DUPLICATES: ClassVar[bool] = False
 
 
 @export
@@ -526,9 +306,10 @@ class TechnologyMapping(Section):
 
 	Used by Vivado command :class:`~pyEDAA.OutputFilter.Xilinx.Commands.SynthesizeDesign`.
 	"""
-	_NAME:   ClassVar[str] = "Technology Mapping"
-	_START:  ClassVar[str] = "Start Technology Mapping"
-	_FINISH: ClassVar[str] = "Finished Technology Mapping : "
+	_NAME:       ClassVar[str] = "Technology Mapping"
+	_START:      ClassVar[str] = "Start Technology Mapping"
+	_FINISH:     ClassVar[str] = "Finished Technology Mapping : "
+	_DUPLICATES: ClassVar[bool] = False
 
 
 @export
@@ -541,6 +322,7 @@ class FlatteningBeforeIOInsertion(SubSection):
 	_NAME:   ClassVar[str] = "Flattening Before IO Insertion"
 	_START:  ClassVar[str] = "Start Flattening Before IO Insertion"
 	_FINISH: ClassVar[str] = "Finished Flattening Before IO Insertion"
+	_DUPLICATES: ClassVar[bool] = False
 
 
 @export
@@ -553,6 +335,7 @@ class FinalNetlistCleanup(SubSection):
 	_NAME:   ClassVar[str] = "Final Netlist Cleanup"
 	_START:  ClassVar[str] = "Start Final Netlist Cleanup"
 	_FINISH: ClassVar[str] = "Finished Final Netlist Cleanup"
+	_DUPLICATES: ClassVar[bool] = False
 
 
 @export
@@ -562,12 +345,13 @@ class IOInsertion(SectionWithChildren):
 
 	Used by Vivado command :class:`~pyEDAA.OutputFilter.Xilinx.Commands.SynthesizeDesign`.
 	"""
-	_NAME:   ClassVar[str] = "IO Insertion"
-	_START:  ClassVar[str] = "Start IO Insertion"
-	_FINISH: ClassVar[str] = "Finished IO Insertion : "
+	_NAME:       ClassVar[str] = "IO Insertion"
+	_START:      ClassVar[str] = "Start IO Insertion"
+	_FINISH:     ClassVar[str] = "Finished IO Insertion : "
+	_DUPLICATES: ClassVar[bool] = False
 
 	# TODO: generalize, use _PARSERS and move to SectionWithChildren
-	def Generator(self, line: Line) -> Generator[Line, Line, Line]:
+	def Generator(self, line: VivadoLine) -> Generator[VivadoLine, VivadoLine, VivadoLine]:
 		line = yield from self._SectionStart(line)
 
 		while True:
@@ -624,9 +408,10 @@ class RenamingGeneratedInstances(Section):
 
 	Used by Vivado command :class:`~pyEDAA.OutputFilter.Xilinx.Commands.SynthesizeDesign`.
 	"""
-	_NAME:   ClassVar[str] = "Renaming Generated Instances"
-	_START:  ClassVar[str] = "Start Renaming Generated Instances"
-	_FINISH: ClassVar[str] = "Finished Renaming Generated Instances : "
+	_NAME:       ClassVar[str] = "Renaming Generated Instances"
+	_START:      ClassVar[str] = "Start Renaming Generated Instances"
+	_FINISH:     ClassVar[str] = "Finished Renaming Generated Instances : "
+	_DUPLICATES: ClassVar[bool] = False
 
 
 @export
@@ -636,9 +421,10 @@ class RebuildingUserHierarchy(Section):
 
 	Used by Vivado command :class:`~pyEDAA.OutputFilter.Xilinx.Commands.SynthesizeDesign`.
 	"""
-	_NAME:   ClassVar[str] = "Rebuilding User Hierarchy"
-	_START:  ClassVar[str] = "Start Rebuilding User Hierarchy"
-	_FINISH: ClassVar[str] = "Finished Rebuilding User Hierarchy : "
+	_NAME:       ClassVar[str] = "Rebuilding User Hierarchy"
+	_START:      ClassVar[str] = "Start Rebuilding User Hierarchy"
+	_FINISH:     ClassVar[str] = "Finished Rebuilding User Hierarchy : "
+	_DUPLICATES: ClassVar[bool] = False
 
 
 @export
@@ -648,9 +434,10 @@ class RenamingGeneratedPorts(Section):
 
 	Used by Vivado command :class:`~pyEDAA.OutputFilter.Xilinx.Commands.SynthesizeDesign`.
 	"""
-	_NAME:   ClassVar[str] = "Renaming Generated Ports"
-	_START:  ClassVar[str] = "Start Renaming Generated Ports"
-	_FINISH: ClassVar[str] = "Finished Renaming Generated Ports : "
+	_NAME:       ClassVar[str] = "Renaming Generated Ports"
+	_START:      ClassVar[str] = "Start Renaming Generated Ports"
+	_FINISH:     ClassVar[str] = "Finished Renaming Generated Ports : "
+	_DUPLICATES: ClassVar[bool] = False
 
 
 @export
@@ -660,9 +447,10 @@ class RenamingGeneratedNets(Section):
 
 	Used by Vivado command :class:`~pyEDAA.OutputFilter.Xilinx.Commands.SynthesizeDesign`.
 	"""
-	_NAME:   ClassVar[str] = "Renaming Generated Nets"
-	_START:  ClassVar[str] = "Start Renaming Generated Nets"
-	_FINISH: ClassVar[str] = "Finished Renaming Generated Nets : "
+	_NAME:       ClassVar[str] = "Renaming Generated Nets"
+	_START:      ClassVar[str] = "Start Renaming Generated Nets"
+	_FINISH:     ClassVar[str] = "Finished Renaming Generated Nets : "
+	_DUPLICATES: ClassVar[bool] = False
 
 
 @export
@@ -672,9 +460,10 @@ class WritingSynthesisReport(Section):
 
 	Used by Vivado command :class:`~pyEDAA.OutputFilter.Xilinx.Commands.SynthesizeDesign`.
 	"""
-	_NAME:   ClassVar[str] = "Writing Synthesis Report"
-	_START:  ClassVar[str] = "Start Writing Synthesis Report"
-	_FINISH: ClassVar[str] = "Finished Writing Synthesis Report : "
+	_NAME:       ClassVar[str] = "Writing Synthesis Report"
+	_START:      ClassVar[str] = "Start Writing Synthesis Report"
+	_FINISH:     ClassVar[str] = "Finished Writing Synthesis Report : "
+	_DUPLICATES: ClassVar[bool] = False
 
 	_blackboxes: Dict[str, int]  #: Blackbox statistics: blackbox name -> count
 	_cells:      Dict[str, int]  #: Cell statistics: cell name -> count
@@ -703,7 +492,7 @@ class WritingSynthesisReport(Section):
 		"""
 		return self._blackboxes
 
-	def _BlackboxesGenerator(self, line: Line) -> Generator[Line, Line, Line]:
+	def _BlackboxesGenerator(self, line: VivadoLine) -> Generator[VivadoLine, VivadoLine, VivadoLine]:
 		"""
 		A parser parsing the blackboxes table.
 
@@ -757,7 +546,7 @@ class WritingSynthesisReport(Section):
 		nextLine = yield line
 		return nextLine
 
-	def _CellGenerator(self, line: Line) -> Generator[Line, Line, Line]:
+	def _CellGenerator(self, line: VivadoLine) -> Generator[VivadoLine, VivadoLine, VivadoLine]:
 		"""
 		A parser parsing the cell statistic table.
 
@@ -811,7 +600,7 @@ class WritingSynthesisReport(Section):
 		nextLine = yield line
 		return nextLine
 
-	def Generator(self, line: Line) -> Generator[Line, Line, Line]:
+	def Generator(self, line: VivadoLine) -> Generator[VivadoLine, VivadoLine, VivadoLine]:
 		line = yield from self._SectionStart(line)
 
 		while True:
