+++ tests/unit/Vivado/Logfiles.py
@@ -33,13 +33,12 @@ from datetime import datetime
 from pathlib  import Path
 from unittest import TestCase as TestCase
 
-from pyTooling.Versioning                        import YearReleaseVersion
-from pyTooling.Warning                           import WarningCollector
+from pyTooling.Versioning       import YearReleaseVersion
+from pyTooling.Warning          import WarningCollector
 
-from pyEDAA.OutputFilter.Xilinx                  import Document, LinkDesign, OptimizeDesign, PlaceDesign, RouteDesign, Line
-from pyEDAA.OutputFilter.Xilinx                  import PhysicalOptimizeDesign, WriteBitstream
-from pyEDAA.OutputFilter.Xilinx.Commands         import SynthesizeDesign
-from pyEDAA.OutputFilter.Xilinx.SynthesizeDesign import WritingSynthesisReport, CrossBoundaryAndAreaOptimization, RTLElaboration
+from pyEDAA.OutputFilter.Xilinx import Document, VivadoLine
+from pyEDAA.OutputFilter.Xilinx import Synth_Design, Link_Design, Opt_Design, Place_Design, Route_Design, PhyOpt_Design, Write_Bitstream
+from pyEDAA.OutputFilter.Xilinx import SynthesizeDesign as _SynthDesign
 
 if __name__ == "__main__": # pragma: no cover
 	print("ERROR: you called a testcase declaration file as an executable module.")
@@ -83,25 +82,122 @@ class Stopwatch(TestCase):
 
 		self.assertEqual(YearReleaseVersion(2025, 1), processor._preamble.ToolVersion)
 
-		synthesis = processor[SynthesizeDesign]
+		synthesis = processor[Synth_Design]
 		self.assertEqual(66, len(synthesis.InfoMessages))
 		self.assertEqual(3, len(synthesis.WarningMessages))
 		self.assertEqual(0, len(synthesis.CriticalWarningMessages))
 		self.assertEqual(0, len(synthesis.ErrorMessages))
 
-		rtlElaboration = synthesis[RTLElaboration]
+		rtlElaboration = synthesis[_SynthDesign.RTLElaboration]
 		self.assertEqual(47, len(rtlElaboration.InfoMessages))
 		self.assertEqual(0, len(rtlElaboration.WarningMessages))
 		self.assertEqual(0, len(rtlElaboration.CriticalWarningMessages))
 		self.assertEqual(0, len(rtlElaboration.ErrorMessages))
 
-		crossBoundaryAndAreaOptimization = synthesis[CrossBoundaryAndAreaOptimization]
+		handlingCustomAttributes = synthesis[_SynthDesign.HandlingCustomAttributes]
+		self.assertEqual(0, len(handlingCustomAttributes.InfoMessages))
+		self.assertEqual(0, len(handlingCustomAttributes.WarningMessages))
+		self.assertEqual(0, len(handlingCustomAttributes.CriticalWarningMessages))
+		self.assertEqual(0, len(handlingCustomAttributes.ErrorMessages))
+
+		loadingPart = synthesis[_SynthDesign.LoadingPart]
+		self.assertEqual(0, len(loadingPart.InfoMessages))
+		self.assertEqual(0, len(loadingPart.WarningMessages))
+		self.assertEqual(0, len(loadingPart.CriticalWarningMessages))
+		self.assertEqual(0, len(loadingPart.ErrorMessages))
+
+		applySetPropertyXDCConstraints = synthesis[_SynthDesign.ApplySetPropertyXDCConstraints]
+		self.assertEqual(0, len(applySetPropertyXDCConstraints.InfoMessages))
+		self.assertEqual(0, len(applySetPropertyXDCConstraints.WarningMessages))
+		self.assertEqual(0, len(applySetPropertyXDCConstraints.CriticalWarningMessages))
+		self.assertEqual(0, len(applySetPropertyXDCConstraints.ErrorMessages))
+
+		rtlComponentStatistics = synthesis[_SynthDesign.RTLComponentStatistics]
+		self.assertEqual(0, len(rtlComponentStatistics.InfoMessages))
+		self.assertEqual(0, len(rtlComponentStatistics.WarningMessages))
+		self.assertEqual(0, len(rtlComponentStatistics.CriticalWarningMessages))
+		self.assertEqual(0, len(rtlComponentStatistics.ErrorMessages))
+
+		partResourceSummary = synthesis[_SynthDesign.PartResourceSummary]
+		self.assertEqual(0, len(partResourceSummary.InfoMessages))
+		self.assertEqual(0, len(partResourceSummary.WarningMessages))
+		self.assertEqual(0, len(partResourceSummary.CriticalWarningMessages))
+		self.assertEqual(0, len(partResourceSummary.ErrorMessages))
+
+		crossBoundaryAndAreaOptimization = synthesis[_SynthDesign.CrossBoundaryAndAreaOptimization]
 		self.assertEqual(0, len(crossBoundaryAndAreaOptimization.InfoMessages))
 		self.assertEqual(3, len(crossBoundaryAndAreaOptimization.WarningMessages))
 		self.assertEqual(0, len(crossBoundaryAndAreaOptimization.CriticalWarningMessages))
 		self.assertEqual(0, len(crossBoundaryAndAreaOptimization.ErrorMessages))
 
-		writingSynthesisReport = synthesis[WritingSynthesisReport]
+		applyingXDCTimingConstraints = synthesis[_SynthDesign.ApplyingXDCTimingConstraints]
+		self.assertEqual(0, len(applyingXDCTimingConstraints.InfoMessages))
+		self.assertEqual(0, len(applyingXDCTimingConstraints.WarningMessages))
+		self.assertEqual(0, len(applyingXDCTimingConstraints.CriticalWarningMessages))
+		self.assertEqual(0, len(applyingXDCTimingConstraints.ErrorMessages))
+
+		timingOptimization = synthesis[_SynthDesign.TimingOptimization]
+		self.assertEqual(0, len(timingOptimization.InfoMessages))
+		self.assertEqual(0, len(timingOptimization.WarningMessages))
+		self.assertEqual(0, len(timingOptimization.CriticalWarningMessages))
+		self.assertEqual(0, len(timingOptimization.ErrorMessages))
+
+		technologyMapping = synthesis[_SynthDesign.TechnologyMapping]
+		self.assertEqual(0, len(technologyMapping.InfoMessages))
+		self.assertEqual(0, len(technologyMapping.WarningMessages))
+		self.assertEqual(0, len(technologyMapping.CriticalWarningMessages))
+		self.assertEqual(0, len(technologyMapping.ErrorMessages))
+
+		ioInsertion = synthesis[_SynthDesign.IOInsertion]
+		self.assertEqual(0, len(ioInsertion.InfoMessages))
+		self.assertEqual(0, len(ioInsertion.WarningMessages))
+		self.assertEqual(0, len(ioInsertion.CriticalWarningMessages))
+		self.assertEqual(0, len(ioInsertion.ErrorMessages))
+
+		flatteningBeforeIOInsertion = ioInsertion[_SynthDesign.FlatteningBeforeIOInsertion]
+		self.assertEqual(0, len(flatteningBeforeIOInsertion.InfoMessages))
+		self.assertEqual(0, len(flatteningBeforeIOInsertion.WarningMessages))
+		self.assertEqual(0, len(flatteningBeforeIOInsertion.CriticalWarningMessages))
+		self.assertEqual(0, len(flatteningBeforeIOInsertion.ErrorMessages))
+
+		finalNetlistCleanup = ioInsertion[_SynthDesign.FinalNetlistCleanup]
+		self.assertEqual(0, len(finalNetlistCleanup.InfoMessages))
+		self.assertEqual(0, len(finalNetlistCleanup.WarningMessages))
+		self.assertEqual(0, len(finalNetlistCleanup.CriticalWarningMessages))
+		self.assertEqual(0, len(finalNetlistCleanup.ErrorMessages))
+
+		renamingGeneratedInstances = synthesis[_SynthDesign.RenamingGeneratedInstances]
+		self.assertEqual(0, len(renamingGeneratedInstances.InfoMessages))
+		self.assertEqual(0, len(renamingGeneratedInstances.WarningMessages))
+		self.assertEqual(0, len(renamingGeneratedInstances.CriticalWarningMessages))
+		self.assertEqual(0, len(renamingGeneratedInstances.ErrorMessages))
+
+		rebuildingUserHierarchy = synthesis[_SynthDesign.RebuildingUserHierarchy]
+		self.assertEqual(0, len(rebuildingUserHierarchy.InfoMessages))
+		self.assertEqual(0, len(rebuildingUserHierarchy.WarningMessages))
+		self.assertEqual(0, len(rebuildingUserHierarchy.CriticalWarningMessages))
+		self.assertEqual(0, len(rebuildingUserHierarchy.ErrorMessages))
+
+		renamingGeneratedPorts = synthesis[_SynthDesign.RenamingGeneratedPorts]
+		self.assertEqual(0, len(renamingGeneratedPorts.InfoMessages))
+		self.assertEqual(0, len(renamingGeneratedPorts.WarningMessages))
+		self.assertEqual(0, len(renamingGeneratedPorts.CriticalWarningMessages))
+		self.assertEqual(0, len(renamingGeneratedPorts.ErrorMessages))
+
+		handlingCustomAttributes = handlingCustomAttributes.Next
+		self.assertIsNotNone(handlingCustomAttributes)
+		self.assertEqual(0, len(handlingCustomAttributes.InfoMessages))
+		self.assertEqual(0, len(handlingCustomAttributes.WarningMessages))
+		self.assertEqual(0, len(handlingCustomAttributes.CriticalWarningMessages))
+		self.assertEqual(0, len(handlingCustomAttributes.ErrorMessages))
+
+		renamingGeneratedNets = synthesis[_SynthDesign.RenamingGeneratedNets]
+		self.assertEqual(0, len(renamingGeneratedNets.InfoMessages))
+		self.assertEqual(0, len(renamingGeneratedNets.WarningMessages))
+		self.assertEqual(0, len(renamingGeneratedNets.CriticalWarningMessages))
+		self.assertEqual(0, len(renamingGeneratedNets.ErrorMessages))
+
+		writingSynthesisReport = synthesis[_SynthDesign.WritingSynthesisReport]
 		self.assertEqual(0, len(writingSynthesisReport.Blackboxes))
 		self.assertEqual(14, len(writingSynthesisReport.Cells))
 
@@ -109,7 +205,7 @@ class Stopwatch(TestCase):
 		lineCount = len(processor.Lines)
 		previousLine = None
 		for i, line in enumerate(processor.Lines, start=1):
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 			self.assertEqual(i, line.LineNumber)
 			self.assertIs(previousLine, line.PreviousLine)
 			if i < lineCount:
@@ -142,37 +238,37 @@ class Stopwatch(TestCase):
 		sumWarn = Aggregator()
 		sumCrit = Aggregator()
 		sumErro = Aggregator()
-		linkDesign = processor[LinkDesign]
+		linkDesign = processor[Link_Design]
 		self.assertEqual(9, sumInfo.sum(len(linkDesign.InfoMessages)))
 		self.assertEqual(2, sumWarn.sum(len(linkDesign.WarningMessages)))
 		self.assertEqual(2, sumCrit.sum(len(linkDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(linkDesign.ErrorMessages)))
 
-		optDesign = processor[OptimizeDesign]
+		optDesign = processor[Opt_Design]
 		self.assertEqual(25, sumInfo.sum(len(optDesign.InfoMessages)))
 		self.assertEqual(0,  sumWarn.sum(len(optDesign.WarningMessages)))
 		self.assertEqual(0,  sumCrit.sum(len(optDesign.CriticalWarningMessages)))
 		self.assertEqual(0,  sumErro.sum(len(optDesign.ErrorMessages)))
 
-		placeDesign = processor[PlaceDesign]
+		placeDesign = processor[Place_Design]
 		self.assertEqual(37, sumInfo.sum(len(placeDesign.InfoMessages)))
 		self.assertEqual(0,  sumWarn.sum(len(placeDesign.WarningMessages)))
 		self.assertEqual(0,  sumCrit.sum(len(placeDesign.CriticalWarningMessages)))
 		self.assertEqual(0,  sumErro.sum(len(placeDesign.ErrorMessages)))
 
-		physOptDesign = processor[PhysicalOptimizeDesign]
+		physOptDesign = processor[PhyOpt_Design]
 		self.assertEqual(22, sumInfo.sum(len(physOptDesign.InfoMessages)))
 		self.assertEqual(0,  sumWarn.sum(len(physOptDesign.WarningMessages)))
 		self.assertEqual(0,  sumCrit.sum(len(physOptDesign.CriticalWarningMessages)))
 		self.assertEqual(0,  sumErro.sum(len(physOptDesign.ErrorMessages)))
 
-		routeDesign = processor[RouteDesign]
+		routeDesign = processor[Route_Design]
 		self.assertEqual(10, sumInfo.sum(len(routeDesign.InfoMessages)))
 		self.assertEqual(0,  sumWarn.sum(len(routeDesign.WarningMessages)))
 		self.assertEqual(0,  sumCrit.sum(len(routeDesign.CriticalWarningMessages)))
 		self.assertEqual(0,  sumErro.sum(len(routeDesign.ErrorMessages)))
 
-		writeBitstream = processor[WriteBitstream]
+		writeBitstream = processor[Write_Bitstream]
 		self.assertEqual(9, sumInfo.sum(len(writeBitstream.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(writeBitstream.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(writeBitstream.CriticalWarningMessages)))
@@ -188,7 +284,7 @@ class Stopwatch(TestCase):
 		lineCount = len(processor.Lines)
 		previousLine = None
 		for i, line in enumerate(processor.Lines, start=1):
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 			self.assertEqual(i, line.LineNumber)
 			self.assertIs(previousLine, line.PreviousLine)
 			if i < lineCount:
@@ -197,6 +293,10 @@ class Stopwatch(TestCase):
 				self.assertIsNone(line.NextLine)
 
 			previousLine = line
+
+
+# class TestCase:
+# 	pass
 
 
 class CERN_DevKit(TestCase):
@@ -217,12 +317,12 @@ class CERN_DevKit(TestCase):
 
 		self.assertEqual(YearReleaseVersion(2024, 2), processor._preamble.ToolVersion)
 
-		synthesis = processor[SynthesizeDesign]
-		self.assertEqual(13, len(synthesis[WritingSynthesisReport].Blackboxes))
+		synthesis = processor[Synth_Design]
+		self.assertEqual(13, len(synthesis[_SynthDesign.WritingSynthesisReport].Blackboxes))
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_ImplementationLogfile(self) -> None:
 		print()
@@ -246,37 +346,37 @@ class CERN_DevKit(TestCase):
 		sumCrit = Aggregator()
 		sumErro = Aggregator()
 
-		linkDesign = processor[LinkDesign]
+		linkDesign = processor[Link_Design]
 		self.assertEqual(30, sumInfo.sum(len(linkDesign.InfoMessages)))
 		self.assertEqual(117, sumWarn.sum(len(linkDesign.WarningMessages)))  # -14
 		self.assertEqual(0, sumCrit.sum(len(linkDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(linkDesign.ErrorMessages)))
 
-		optDesign = processor[OptimizeDesign]
+		optDesign = processor[Opt_Design]
 		self.assertEqual(145, sumInfo.sum(len(optDesign.InfoMessages)))
 		self.assertEqual(6,  sumWarn.sum(len(optDesign.WarningMessages)))
 		self.assertEqual(0,  sumCrit.sum(len(optDesign.CriticalWarningMessages)))
 		self.assertEqual(0,  sumErro.sum(len(optDesign.ErrorMessages)))
 
-		placeDesign = processor[PlaceDesign]
+		placeDesign = processor[Place_Design]
 		self.assertEqual(59, sumInfo.sum(len(placeDesign.InfoMessages)))
 		self.assertEqual(0,  sumWarn.sum(len(placeDesign.WarningMessages)))
 		self.assertEqual(1,  sumCrit.sum(len(placeDesign.CriticalWarningMessages)))
 		self.assertEqual(0,  sumErro.sum(len(placeDesign.ErrorMessages)))
 
-		physOptDesign = processor[PhysicalOptimizeDesign]
+		physOptDesign = processor[PhyOpt_Design]
 		self.assertEqual(6, sumInfo.sum(len(physOptDesign.InfoMessages)))
 		self.assertEqual(0,  sumWarn.sum(len(physOptDesign.WarningMessages)))
 		self.assertEqual(0,  sumCrit.sum(len(physOptDesign.CriticalWarningMessages)))
 		self.assertEqual(0,  sumErro.sum(len(physOptDesign.ErrorMessages)))
 
-		routeDesign = processor[RouteDesign]
+		routeDesign = processor[Route_Design]
 		self.assertEqual(13, sumInfo.sum(len(routeDesign.InfoMessages)))
 		self.assertEqual(0,  sumWarn.sum(len(routeDesign.WarningMessages)))
 		self.assertEqual(0,  sumCrit.sum(len(routeDesign.CriticalWarningMessages)))
 		self.assertEqual(0,  sumErro.sum(len(routeDesign.ErrorMessages)))
 
-		writeBitstream = processor[WriteBitstream]
+		writeBitstream = processor[Write_Bitstream]
 		self.assertEqual(9, sumInfo.sum(len(writeBitstream.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(writeBitstream.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(writeBitstream.CriticalWarningMessages)))
@@ -290,7 +390,7 @@ class CERN_DevKit(TestCase):
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 
 class Enclustra_Mercury_ZX5(TestCase):
@@ -316,13 +416,13 @@ class Enclustra_Mercury_ZX5(TestCase):
 		sumCrit = Aggregator()
 		sumErro = Aggregator()
 
-		synthesis = processor[SynthesizeDesign]
+		synthesis = processor[Synth_Design]
 		self.assertEqual(65, sumInfo.sum(len(synthesis.InfoMessages)))      # -12
 		self.assertEqual(111, sumWarn.sum(len(synthesis.WarningMessages)))  # -2
 		self.assertEqual(0, sumCrit.sum(len(synthesis.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(synthesis.ErrorMessages)))
 
-		self.assertEqual(14, len(synthesis[WritingSynthesisReport].Blackboxes))
+		self.assertEqual(14, len(synthesis[_SynthDesign.WritingSynthesisReport].Blackboxes))
 
 		# compare sum of sections with total numbers
 		self.assertGreaterEqual(len(processor.InfoMessages), sumInfo.Value)
@@ -332,7 +432,7 @@ class Enclustra_Mercury_ZX5(TestCase):
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_ImplementationLogfile_2019_1(self) -> None:
 		print()
@@ -355,33 +455,33 @@ class Enclustra_Mercury_ZX5(TestCase):
 		sumWarn = Aggregator()
 		sumCrit = Aggregator()
 		sumErro = Aggregator()
-		linkDesign = processor[LinkDesign]
+		linkDesign = processor[Link_Design]
 		self.assertEqual(21, sumInfo.sum(len(linkDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(linkDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(linkDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(linkDesign.ErrorMessages)))
 
-		optDesign = processor[OptimizeDesign]
+		optDesign = processor[Opt_Design]
 		self.assertEqual(24, sumInfo.sum(len(optDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(optDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(optDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(optDesign.ErrorMessages)))
 
-		placeDesign = processor[PlaceDesign]
+		placeDesign = processor[Place_Design]
 		self.assertEqual(21, sumInfo.sum(len(placeDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(placeDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(placeDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(placeDesign.ErrorMessages)))
 
-		self.assertNotIn(PhysicalOptimizeDesign, processor)
+		self.assertNotIn(PhyOpt_Design, processor)
 
-		routeDesign = processor[RouteDesign]
+		routeDesign = processor[Route_Design]
 		self.assertEqual(12, sumInfo.sum(len(routeDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(routeDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(routeDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(routeDesign.ErrorMessages)))
 
-		self.assertNotIn(WriteBitstream, processor)
+		self.assertNotIn(Write_Bitstream, processor)
 
 		# compare sum of sections with total numbers
 		self.assertGreaterEqual(len(processor.InfoMessages), sumInfo.Value)
@@ -391,7 +491,7 @@ class Enclustra_Mercury_ZX5(TestCase):
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_SynthesisLogfile_2019_2(self) -> None:
 		print()
@@ -410,12 +510,12 @@ class Enclustra_Mercury_ZX5(TestCase):
 
 		self.assertEqual(YearReleaseVersion(2019, 2), processor._preamble.ToolVersion)
 
-		synthesis = processor[SynthesizeDesign]
-		self.assertEqual(14, len(synthesis[WritingSynthesisReport].Blackboxes))
+		synthesis = processor[Synth_Design]
+		self.assertEqual(14, len(synthesis[_SynthDesign.WritingSynthesisReport].Blackboxes))
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_ImplementationLogfile_2019_2(self) -> None:
 		print()
@@ -438,33 +538,33 @@ class Enclustra_Mercury_ZX5(TestCase):
 		sumWarn = Aggregator()
 		sumCrit = Aggregator()
 		sumErro = Aggregator()
-		linkDesign = processor[LinkDesign]
+		linkDesign = processor[Link_Design]
 		self.assertEqual(21, sumInfo.sum(len(linkDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(linkDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(linkDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(linkDesign.ErrorMessages)))
 
-		optDesign = processor[OptimizeDesign]
+		optDesign = processor[Opt_Design]
 		self.assertEqual(27, sumInfo.sum(len(optDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(optDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(optDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(optDesign.ErrorMessages)))
 
-		placeDesign = processor[PlaceDesign]
+		placeDesign = processor[Place_Design]
 		self.assertEqual(24, sumInfo.sum(len(placeDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(placeDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(placeDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(placeDesign.ErrorMessages)))
 
-		self.assertNotIn(PhysicalOptimizeDesign, processor)
+		self.assertNotIn(PhyOpt_Design, processor)
 
-		routeDesign = processor[RouteDesign]
+		routeDesign = processor[Route_Design]
 		self.assertEqual(13, sumInfo.sum(len(routeDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(routeDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(routeDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(routeDesign.ErrorMessages)))
 
-		writeBitstream = processor[WriteBitstream]
+		writeBitstream = processor[Write_Bitstream]
 		self.assertEqual(9, sumInfo.sum(len(writeBitstream.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(writeBitstream.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(writeBitstream.CriticalWarningMessages)))
@@ -478,7 +578,7 @@ class Enclustra_Mercury_ZX5(TestCase):
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_SynthesisLogfile_2020_1(self) -> None:
 		print()
@@ -497,12 +597,12 @@ class Enclustra_Mercury_ZX5(TestCase):
 
 		self.assertEqual(YearReleaseVersion(2020, 1), processor._preamble.ToolVersion)
 
-		synthesis = processor[SynthesizeDesign]
-		self.assertEqual(7, len(synthesis[WritingSynthesisReport].Blackboxes))
+		synthesis = processor[Synth_Design]
+		self.assertEqual(7, len(synthesis[_SynthDesign.WritingSynthesisReport].Blackboxes))
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_ImplementationLogfile_2020_1(self) -> None:
 		print()
@@ -525,37 +625,37 @@ class Enclustra_Mercury_ZX5(TestCase):
 		sumWarn = Aggregator()
 		sumCrit = Aggregator()
 		sumErro = Aggregator()
-		linkDesign = processor[LinkDesign]
+		linkDesign = processor[Link_Design]
 		self.assertEqual(14, sumInfo.sum(len(linkDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(linkDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(linkDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(linkDesign.ErrorMessages)))
 
-		optDesign = processor[OptimizeDesign]
+		optDesign = processor[Opt_Design]
 		self.assertEqual(26, sumInfo.sum(len(optDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(optDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(optDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(optDesign.ErrorMessages)))
 
-		placeDesign = processor[PlaceDesign]
+		placeDesign = processor[Place_Design]
 		self.assertEqual(28, sumInfo.sum(len(placeDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(placeDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(placeDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(placeDesign.ErrorMessages)))
 
-		physOptDesign = processor[PhysicalOptimizeDesign]
+		physOptDesign = processor[PhyOpt_Design]
 		self.assertEqual(4, sumInfo.sum(len(physOptDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(physOptDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(physOptDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(physOptDesign.ErrorMessages)))
 
-		routeDesign = processor[RouteDesign]
+		routeDesign = processor[Route_Design]
 		self.assertEqual(13, sumInfo.sum(len(routeDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(routeDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(routeDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(routeDesign.ErrorMessages)))
 
-		writeBitstream = processor[WriteBitstream]
+		writeBitstream = processor[Write_Bitstream]
 		self.assertEqual(8, sumInfo.sum(len(writeBitstream.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(writeBitstream.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(writeBitstream.CriticalWarningMessages)))
@@ -569,7 +669,7 @@ class Enclustra_Mercury_ZX5(TestCase):
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_SynthesisLogfile_2020_2(self) -> None:
 		print()
@@ -588,12 +688,12 @@ class Enclustra_Mercury_ZX5(TestCase):
 
 		self.assertEqual(YearReleaseVersion(2020, 2), processor._preamble.ToolVersion)
 
-		synthesis = processor[SynthesizeDesign]
-		self.assertEqual(7, len(synthesis[WritingSynthesisReport].Blackboxes))
+		synthesis = processor[Synth_Design]
+		self.assertEqual(7, len(synthesis[_SynthDesign.WritingSynthesisReport].Blackboxes))
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_ImplementationLogfile_2020_2(self) -> None:
 		print()
@@ -616,37 +716,37 @@ class Enclustra_Mercury_ZX5(TestCase):
 		sumWarn = Aggregator()
 		sumCrit = Aggregator()
 		sumErro = Aggregator()
-		linkDesign = processor[LinkDesign]
+		linkDesign = processor[Link_Design]
 		self.assertEqual(15, sumInfo.sum(len(linkDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(linkDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(linkDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(linkDesign.ErrorMessages)))
 
-		optDesign = processor[OptimizeDesign]
+		optDesign = processor[Opt_Design]
 		self.assertEqual(26, sumInfo.sum(len(optDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(optDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(optDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(optDesign.ErrorMessages)))
 
-		placeDesign = processor[PlaceDesign]
+		placeDesign = processor[Place_Design]
 		self.assertEqual(28, sumInfo.sum(len(placeDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(placeDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(placeDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(placeDesign.ErrorMessages)))
 
-		physOptDesign = processor[PhysicalOptimizeDesign]
+		physOptDesign = processor[PhyOpt_Design]
 		self.assertEqual(4, sumInfo.sum(len(physOptDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(physOptDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(physOptDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(physOptDesign.ErrorMessages)))
 
-		routeDesign = processor[RouteDesign]
+		routeDesign = processor[Route_Design]
 		self.assertEqual(12, sumInfo.sum(len(routeDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(routeDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(routeDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(routeDesign.ErrorMessages)))
 
-		writeBitstream = processor[WriteBitstream]
+		writeBitstream = processor[Write_Bitstream]
 		self.assertEqual(8, sumInfo.sum(len(writeBitstream.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(writeBitstream.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(writeBitstream.CriticalWarningMessages)))
@@ -660,7 +760,7 @@ class Enclustra_Mercury_ZX5(TestCase):
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_SynthesisLogfile_2021_1(self) -> None:
 		print()
@@ -679,8 +779,8 @@ class Enclustra_Mercury_ZX5(TestCase):
 
 		self.assertEqual(YearReleaseVersion(2021, 1), processor._preamble.ToolVersion)
 
-		synthesis = processor[SynthesizeDesign]
-		self.assertEqual(7, len(synthesis[WritingSynthesisReport].Blackboxes))
+		synthesis = processor[Synth_Design]
+		self.assertEqual(7, len(synthesis[_SynthDesign.WritingSynthesisReport].Blackboxes))
 
 	def test_ImplementationLogfile_2021_1(self) -> None:
 		print()
@@ -703,37 +803,37 @@ class Enclustra_Mercury_ZX5(TestCase):
 		sumWarn = Aggregator()
 		sumCrit = Aggregator()
 		sumErro = Aggregator()
-		linkDesign = processor[LinkDesign]
+		linkDesign = processor[Link_Design]
 		self.assertEqual(15, sumInfo.sum(len(linkDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(linkDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(linkDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(linkDesign.ErrorMessages)))
 
-		optDesign = processor[OptimizeDesign]
+		optDesign = processor[Opt_Design]
 		self.assertEqual(26, sumInfo.sum(len(optDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(optDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(optDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(optDesign.ErrorMessages)))
 
-		placeDesign = processor[PlaceDesign]
+		placeDesign = processor[Place_Design]
 		self.assertEqual(28, sumInfo.sum(len(placeDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(placeDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(placeDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(placeDesign.ErrorMessages)))
 
-		physOptDesign = processor[PhysicalOptimizeDesign]
+		physOptDesign = processor[PhyOpt_Design]
 		self.assertEqual(4, sumInfo.sum(len(physOptDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(physOptDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(physOptDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(physOptDesign.ErrorMessages)))
 
-		routeDesign = processor[RouteDesign]
+		routeDesign = processor[Route_Design]
 		self.assertEqual(13, sumInfo.sum(len(routeDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(routeDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(routeDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(routeDesign.ErrorMessages)))
 
-		writeBitstream = processor[WriteBitstream]
+		writeBitstream = processor[Write_Bitstream]
 		self.assertEqual(8, sumInfo.sum(len(writeBitstream.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(writeBitstream.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(writeBitstream.CriticalWarningMessages)))
@@ -747,7 +847,7 @@ class Enclustra_Mercury_ZX5(TestCase):
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_SynthesisLogfile_2021_2(self) -> None:
 		print()
@@ -766,12 +866,12 @@ class Enclustra_Mercury_ZX5(TestCase):
 
 		self.assertEqual(YearReleaseVersion(2021, 2), processor._preamble.ToolVersion)
 
-		synthesis = processor[SynthesizeDesign]
-		self.assertEqual(7, len(synthesis[WritingSynthesisReport].Blackboxes))
+		synthesis = processor[Synth_Design]
+		self.assertEqual(7, len(synthesis[_SynthDesign.WritingSynthesisReport].Blackboxes))
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_ImplementationLogfile_2021_2(self) -> None:
 		print()
@@ -794,37 +894,37 @@ class Enclustra_Mercury_ZX5(TestCase):
 		sumWarn = Aggregator()
 		sumCrit = Aggregator()
 		sumErro = Aggregator()
-		linkDesign = processor[LinkDesign]
+		linkDesign = processor[Link_Design]
 		self.assertEqual(15, sumInfo.sum(len(linkDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(linkDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(linkDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(linkDesign.ErrorMessages)))
 
-		optDesign = processor[OptimizeDesign]
+		optDesign = processor[Opt_Design]
 		self.assertEqual(30, sumInfo.sum(len(optDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(optDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(optDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(optDesign.ErrorMessages)))
 
-		placeDesign = processor[PlaceDesign]
+		placeDesign = processor[Place_Design]
 		self.assertEqual(28, sumInfo.sum(len(placeDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(placeDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(placeDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(placeDesign.ErrorMessages)))
 
-		physOptDesign = processor[PhysicalOptimizeDesign]
+		physOptDesign = processor[PhyOpt_Design]
 		self.assertEqual(4, sumInfo.sum(len(physOptDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(physOptDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(physOptDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(physOptDesign.ErrorMessages)))
 
-		routeDesign = processor[RouteDesign]
+		routeDesign = processor[Route_Design]
 		self.assertEqual(13, sumInfo.sum(len(routeDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(routeDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(routeDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(routeDesign.ErrorMessages)))
 
-		writeBitstream = processor[WriteBitstream]
+		writeBitstream = processor[Write_Bitstream]
 		self.assertEqual(8, sumInfo.sum(len(writeBitstream.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(writeBitstream.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(writeBitstream.CriticalWarningMessages)))
@@ -838,7 +938,7 @@ class Enclustra_Mercury_ZX5(TestCase):
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_SynthesisLogfile_2022_1(self) -> None:
 		print()
@@ -857,12 +957,12 @@ class Enclustra_Mercury_ZX5(TestCase):
 
 		self.assertEqual(YearReleaseVersion(2022, 1), processor._preamble.ToolVersion)
 
-		synthesis = processor[SynthesizeDesign]
-		self.assertEqual(7, len(synthesis[WritingSynthesisReport].Blackboxes))
+		synthesis = processor[Synth_Design]
+		self.assertEqual(7, len(synthesis[_SynthDesign.WritingSynthesisReport].Blackboxes))
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_ImplementationLogfile_2022_1(self) -> None:
 		print()
@@ -885,37 +985,37 @@ class Enclustra_Mercury_ZX5(TestCase):
 		sumWarn = Aggregator()
 		sumCrit = Aggregator()
 		sumErro = Aggregator()
-		linkDesign = processor[LinkDesign]
+		linkDesign = processor[Link_Design]
 		self.assertEqual(15, sumInfo.sum(len(linkDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(linkDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(linkDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(linkDesign.ErrorMessages)))
 
-		optDesign = processor[OptimizeDesign]
+		optDesign = processor[Opt_Design]
 		self.assertEqual(28, sumInfo.sum(len(optDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(optDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(optDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(optDesign.ErrorMessages)))
 
-		placeDesign = processor[PlaceDesign]
+		placeDesign = processor[Place_Design]
 		self.assertEqual(28, sumInfo.sum(len(placeDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(placeDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(placeDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(placeDesign.ErrorMessages)))
 
-		physOptDesign = processor[PhysicalOptimizeDesign]
+		physOptDesign = processor[PhyOpt_Design]
 		self.assertEqual(4, sumInfo.sum(len(physOptDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(physOptDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(physOptDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(physOptDesign.ErrorMessages)))
 
-		routeDesign = processor[RouteDesign]
+		routeDesign = processor[Route_Design]
 		self.assertEqual(13, sumInfo.sum(len(routeDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(routeDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(routeDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(routeDesign.ErrorMessages)))
 
-		writeBitstream = processor[WriteBitstream]
+		writeBitstream = processor[Write_Bitstream]
 		self.assertEqual(8, sumInfo.sum(len(writeBitstream.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(writeBitstream.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(writeBitstream.CriticalWarningMessages)))
@@ -929,7 +1029,7 @@ class Enclustra_Mercury_ZX5(TestCase):
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_SynthesisLogfile_2022_2(self) -> None:
 		print()
@@ -948,12 +1048,12 @@ class Enclustra_Mercury_ZX5(TestCase):
 
 		self.assertEqual(YearReleaseVersion(2022, 2), processor._preamble.ToolVersion)
 
-		synthesis = processor[SynthesizeDesign]
-		self.assertEqual(7, len(synthesis[WritingSynthesisReport].Blackboxes))
+		synthesis = processor[Synth_Design]
+		self.assertEqual(7, len(synthesis[_SynthDesign.WritingSynthesisReport].Blackboxes))
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_ImplementationLogfile_2022_2(self) -> None:
 		print()
@@ -976,37 +1076,37 @@ class Enclustra_Mercury_ZX5(TestCase):
 		sumWarn = Aggregator()
 		sumCrit = Aggregator()
 		sumErro = Aggregator()
-		linkDesign = processor[LinkDesign]
+		linkDesign = processor[Link_Design]
 		self.assertEqual(15, sumInfo.sum(len(linkDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(linkDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(linkDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(linkDesign.ErrorMessages)))
 
-		optDesign = processor[OptimizeDesign]
+		optDesign = processor[Opt_Design]
 		self.assertEqual(28, sumInfo.sum(len(optDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(optDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(optDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(optDesign.ErrorMessages)))
 
-		placeDesign = processor[PlaceDesign]
+		placeDesign = processor[Place_Design]
 		self.assertEqual(29, sumInfo.sum(len(placeDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(placeDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(placeDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(placeDesign.ErrorMessages)))
 
-		physOptDesign = processor[PhysicalOptimizeDesign]
+		physOptDesign = processor[PhyOpt_Design]
 		self.assertEqual(4, sumInfo.sum(len(physOptDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(physOptDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(physOptDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(physOptDesign.ErrorMessages)))
 
-		routeDesign = processor[RouteDesign]
+		routeDesign = processor[Route_Design]
 		self.assertEqual(13, sumInfo.sum(len(routeDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(routeDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(routeDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(routeDesign.ErrorMessages)))
 
-		writeBitstream = processor[WriteBitstream]
+		writeBitstream = processor[Write_Bitstream]
 		self.assertEqual(8, sumInfo.sum(len(writeBitstream.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(writeBitstream.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(writeBitstream.CriticalWarningMessages)))
@@ -1020,7 +1120,7 @@ class Enclustra_Mercury_ZX5(TestCase):
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_SynthesisLogfile_2023_1(self) -> None:
 		print()
@@ -1039,12 +1139,12 @@ class Enclustra_Mercury_ZX5(TestCase):
 
 		self.assertEqual(YearReleaseVersion(2023, 1), processor._preamble.ToolVersion)
 
-		synthesis = processor[SynthesizeDesign]
-		self.assertEqual(7, len(synthesis[WritingSynthesisReport].Blackboxes))
+		synthesis = processor[Synth_Design]
+		self.assertEqual(7, len(synthesis[_SynthDesign.WritingSynthesisReport].Blackboxes))
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_ImplementationLogfile_2023_1(self) -> None:
 		print()
@@ -1067,37 +1167,37 @@ class Enclustra_Mercury_ZX5(TestCase):
 		sumWarn = Aggregator()
 		sumCrit = Aggregator()
 		sumErro = Aggregator()
-		linkDesign = processor[LinkDesign]
+		linkDesign = processor[Link_Design]
 		self.assertEqual(15, sumInfo.sum(len(linkDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(linkDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(linkDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(linkDesign.ErrorMessages)))
 
-		optDesign = processor[OptimizeDesign]
+		optDesign = processor[Opt_Design]
 		self.assertEqual(27, sumInfo.sum(len(optDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(optDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(optDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(optDesign.ErrorMessages)))
 
-		placeDesign = processor[PlaceDesign]
+		placeDesign = processor[Place_Design]
 		self.assertEqual(29, sumInfo.sum(len(placeDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(placeDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(placeDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(placeDesign.ErrorMessages)))
 
-		physOptDesign = processor[PhysicalOptimizeDesign]
+		physOptDesign = processor[PhyOpt_Design]
 		self.assertEqual(4, sumInfo.sum(len(physOptDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(physOptDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(physOptDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(physOptDesign.ErrorMessages)))
 
-		routeDesign = processor[RouteDesign]
+		routeDesign = processor[Route_Design]
 		self.assertEqual(13, sumInfo.sum(len(routeDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(routeDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(routeDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(routeDesign.ErrorMessages)))
 
-		writeBitstream = processor[WriteBitstream]
+		writeBitstream = processor[Write_Bitstream]
 		self.assertEqual(8, sumInfo.sum(len(writeBitstream.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(writeBitstream.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(writeBitstream.CriticalWarningMessages)))
@@ -1111,7 +1211,7 @@ class Enclustra_Mercury_ZX5(TestCase):
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_SynthesisLogfile_2023_2(self) -> None:
 		print()
@@ -1132,12 +1232,12 @@ class Enclustra_Mercury_ZX5(TestCase):
 
 		self.assertEqual(YearReleaseVersion(2023, 2), processor._preamble.ToolVersion)
 
-		synthesis = processor[SynthesizeDesign]
-		self.assertEqual(7, len(synthesis[WritingSynthesisReport].Blackboxes))
+		synthesis = processor[Synth_Design]
+		self.assertEqual(7, len(synthesis[_SynthDesign.WritingSynthesisReport].Blackboxes))
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_ImplementationLogfile_2023_2(self) -> None:
 		print()
@@ -1160,37 +1260,37 @@ class Enclustra_Mercury_ZX5(TestCase):
 		sumWarn = Aggregator()
 		sumCrit = Aggregator()
 		sumErro = Aggregator()
-		linkDesign = processor[LinkDesign]
+		linkDesign = processor[Link_Design]
 		self.assertEqual(18, sumInfo.sum(len(linkDesign.InfoMessages)))
 		self.assertEqual(101, sumWarn.sum(len(linkDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(linkDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(linkDesign.ErrorMessages)))
 
-		optDesign = processor[OptimizeDesign]
+		optDesign = processor[Opt_Design]
 		self.assertEqual(29, sumInfo.sum(len(optDesign.InfoMessages)))
 		self.assertEqual(100, sumWarn.sum(len(optDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(optDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(optDesign.ErrorMessages)))
 
-		placeDesign = processor[PlaceDesign]
+		placeDesign = processor[Place_Design]
 		self.assertEqual(29, sumInfo.sum(len(placeDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(placeDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(placeDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(placeDesign.ErrorMessages)))
 
-		physOptDesign = processor[PhysicalOptimizeDesign]
+		physOptDesign = processor[PhyOpt_Design]
 		self.assertEqual(4, sumInfo.sum(len(physOptDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(physOptDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(physOptDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(physOptDesign.ErrorMessages)))
 
-		routeDesign = processor[RouteDesign]
+		routeDesign = processor[Route_Design]
 		self.assertEqual(13, sumInfo.sum(len(routeDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(routeDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(routeDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(routeDesign.ErrorMessages)))
 
-		writeBitstream = processor[WriteBitstream]
+		writeBitstream = processor[Write_Bitstream]
 		self.assertEqual(8, sumInfo.sum(len(writeBitstream.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(writeBitstream.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(writeBitstream.CriticalWarningMessages)))
@@ -1204,7 +1304,7 @@ class Enclustra_Mercury_ZX5(TestCase):
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_SynthesisLogfile_2024_1(self) -> None:
 		print()
@@ -1223,12 +1323,12 @@ class Enclustra_Mercury_ZX5(TestCase):
 
 		self.assertEqual(YearReleaseVersion(2024, 1), processor._preamble.ToolVersion)
 
-		synthesis = processor[SynthesizeDesign]
-		self.assertEqual(7, len(synthesis[WritingSynthesisReport].Blackboxes))
+		synthesis = processor[Synth_Design]
+		self.assertEqual(7, len(synthesis[_SynthDesign.WritingSynthesisReport].Blackboxes))
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_ImplementationLogfile_2024_1(self) -> None:
 		print()
@@ -1251,37 +1351,37 @@ class Enclustra_Mercury_ZX5(TestCase):
 		sumWarn = Aggregator()
 		sumCrit = Aggregator()
 		sumErro = Aggregator()
-		linkDesign = processor[LinkDesign]
+		linkDesign = processor[Link_Design]
 		self.assertEqual(19, sumInfo.sum(len(linkDesign.InfoMessages)))
 		self.assertEqual(101, sumWarn.sum(len(linkDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(linkDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(linkDesign.ErrorMessages)))
 
-		optDesign = processor[OptimizeDesign]
+		optDesign = processor[Opt_Design]
 		self.assertEqual(28, sumInfo.sum(len(optDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(optDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(optDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(optDesign.ErrorMessages)))
 
-		placeDesign = processor[PlaceDesign]
+		placeDesign = processor[Place_Design]
 		self.assertEqual(29, sumInfo.sum(len(placeDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(placeDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(placeDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(placeDesign.ErrorMessages)))
 
-		physOptDesign = processor[PhysicalOptimizeDesign]
+		physOptDesign = processor[PhyOpt_Design]
 		self.assertEqual(5, sumInfo.sum(len(physOptDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(physOptDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(physOptDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(physOptDesign.ErrorMessages)))
 
-		routeDesign = processor[RouteDesign]
+		routeDesign = processor[Route_Design]
 		self.assertEqual(12, sumInfo.sum(len(routeDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(routeDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(routeDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(routeDesign.ErrorMessages)))
 
-		writeBitstream = processor[WriteBitstream]
+		writeBitstream = processor[Write_Bitstream]
 		self.assertEqual(8, sumInfo.sum(len(writeBitstream.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(writeBitstream.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(writeBitstream.CriticalWarningMessages)))
@@ -1295,7 +1395,7 @@ class Enclustra_Mercury_ZX5(TestCase):
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_SynthesisLogfile_2024_2(self) -> None:
 		print()
@@ -1314,12 +1414,12 @@ class Enclustra_Mercury_ZX5(TestCase):
 
 		self.assertEqual(YearReleaseVersion(2024, 2), processor._preamble.ToolVersion)
 
-		synthesis = processor[SynthesizeDesign]
-		self.assertEqual(7, len(synthesis[WritingSynthesisReport].Blackboxes))
+		synthesis = processor[Synth_Design]
+		self.assertEqual(7, len(synthesis[_SynthDesign.WritingSynthesisReport].Blackboxes))
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_ImplementationLogfile_2024_2(self) -> None:
 		print()
@@ -1342,37 +1442,37 @@ class Enclustra_Mercury_ZX5(TestCase):
 		sumWarn = Aggregator()
 		sumCrit = Aggregator()
 		sumErro = Aggregator()
-		linkDesign = processor[LinkDesign]
+		linkDesign = processor[Link_Design]
 		self.assertEqual(15, sumInfo.sum(len(linkDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(linkDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(linkDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(linkDesign.ErrorMessages)))
 
-		optDesign = processor[OptimizeDesign]
+		optDesign = processor[Opt_Design]
 		self.assertEqual(31, sumInfo.sum(len(optDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(optDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(optDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(optDesign.ErrorMessages)))
 
-		placeDesign = processor[PlaceDesign]
+		placeDesign = processor[Place_Design]
 		self.assertEqual(29, sumInfo.sum(len(placeDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(placeDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(placeDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(placeDesign.ErrorMessages)))
 
-		physOptDesign = processor[PhysicalOptimizeDesign]
+		physOptDesign = processor[PhyOpt_Design]
 		self.assertEqual(5, sumInfo.sum(len(physOptDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(physOptDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(physOptDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(physOptDesign.ErrorMessages)))
 
-		routeDesign = processor[RouteDesign]
+		routeDesign = processor[Route_Design]
 		self.assertEqual(10, sumInfo.sum(len(routeDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(routeDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(routeDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(routeDesign.ErrorMessages)))
 
-		writeBitstream = processor[WriteBitstream]
+		writeBitstream = processor[Write_Bitstream]
 		self.assertEqual(8, sumInfo.sum(len(writeBitstream.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(writeBitstream.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(writeBitstream.CriticalWarningMessages)))
@@ -1386,7 +1486,7 @@ class Enclustra_Mercury_ZX5(TestCase):
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_SynthesisLogfile_2025_1(self) -> None:
 		print()
@@ -1405,12 +1505,12 @@ class Enclustra_Mercury_ZX5(TestCase):
 
 		self.assertEqual(YearReleaseVersion(2025, 1), processor._preamble.ToolVersion)
 
-		synthesis = processor[SynthesizeDesign]
-		self.assertEqual(7, len(synthesis[WritingSynthesisReport].Blackboxes))
+		synthesis = processor[Synth_Design]
+		self.assertEqual(7, len(synthesis[_SynthDesign.WritingSynthesisReport].Blackboxes))
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_ImplementationLogfile_2025_1(self) -> None:
 		print()
@@ -1433,37 +1533,37 @@ class Enclustra_Mercury_ZX5(TestCase):
 		sumWarn = Aggregator()
 		sumCrit = Aggregator()
 		sumErro = Aggregator()
-		linkDesign = processor[LinkDesign]
+		linkDesign = processor[Link_Design]
 		self.assertEqual(16, sumInfo.sum(len(linkDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(linkDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(linkDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(linkDesign.ErrorMessages)))
 
-		optDesign = processor[OptimizeDesign]
+		optDesign = processor[Opt_Design]
 		self.assertEqual(31, sumInfo.sum(len(optDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(optDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(optDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(optDesign.ErrorMessages)))
 
-		placeDesign = processor[PlaceDesign]
+		placeDesign = processor[Place_Design]
 		self.assertEqual(29, sumInfo.sum(len(placeDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(placeDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(placeDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(placeDesign.ErrorMessages)))
 
-		physOptDesign = processor[PhysicalOptimizeDesign]
+		physOptDesign = processor[PhyOpt_Design]
 		self.assertEqual(5, sumInfo.sum(len(physOptDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(physOptDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(physOptDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(physOptDesign.ErrorMessages)))
 
-		routeDesign = processor[RouteDesign]
+		routeDesign = processor[Route_Design]
 		self.assertEqual(10, sumInfo.sum(len(routeDesign.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(routeDesign.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(routeDesign.CriticalWarningMessages)))
 		self.assertEqual(0, sumErro.sum(len(routeDesign.ErrorMessages)))
 
-		writeBitstream = processor[WriteBitstream]
+		writeBitstream = processor[Write_Bitstream]
 		self.assertEqual(8, sumInfo.sum(len(writeBitstream.InfoMessages)))
 		self.assertEqual(0, sumWarn.sum(len(writeBitstream.WarningMessages)))
 		self.assertEqual(0, sumCrit.sum(len(writeBitstream.CriticalWarningMessages)))
@@ -1477,4 +1577,4 @@ class Enclustra_Mercury_ZX5(TestCase):
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
