+++ tests/unit/Vivado/Synthesis.py
@@ -37,10 +37,8 @@ from unittest import TestCase as TestCase
 from pyTooling.Versioning import YearReleaseVersion
 from pyTooling.Warning    import WarningCollector
 
-from pyEDAA.OutputFilter.Xilinx                  import Processor, SynthesizeDesign, Line
-from pyEDAA.OutputFilter.Xilinx.SynthesizeDesign import RTLElaboration, LoadingPart, RTLComponentStatistics
-from pyEDAA.OutputFilter.Xilinx.SynthesizeDesign import IOInsertion, FlatteningBeforeIOInsertion, FinalNetlistCleanup
-from pyEDAA.OutputFilter.Xilinx.SynthesizeDesign import WritingSynthesisReport
+from pyEDAA.OutputFilter.Xilinx import Processor, Synth_Design, VivadoLine
+from pyEDAA.OutputFilter.Xilinx import SynthesizeDesign as _SynthDesign
 
 if __name__ == "__main__": # pragma: no cover
 	print("ERROR: you called a testcase declaration file as an executable module.")
@@ -95,8 +93,8 @@ class SynthDesign(TestCase):
 		self.assertEqual(0, len(processor.CriticalWarningMessages))
 		self.assertEqual(0, len(processor.ErrorMessages))
 
-		self.assertIn(SynthesizeDesign, processor)
-		synthDesign = processor[SynthesizeDesign]
+		self.assertIn(Synth_Design, processor)
+		synthDesign = processor[Synth_Design]
 		self.assertEqual(0, len(synthDesign.InfoMessages))
 		self.assertEqual(0, len(synthDesign.WarningMessages))
 		self.assertEqual(0, len(synthDesign.CriticalWarningMessages))
@@ -104,7 +102,7 @@ class SynthDesign(TestCase):
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_RTLElaboration(self) -> None:
 		print()
@@ -138,15 +136,15 @@ class SynthDesign(TestCase):
 		self.assertEqual(0, len(processor.CriticalWarningMessages))
 		self.assertEqual(0, len(processor.ErrorMessages))
 
-		self.assertIn(SynthesizeDesign, processor)
-		synthDesign = processor[SynthesizeDesign]
+		self.assertIn(Synth_Design, processor)
+		synthDesign = processor[Synth_Design]
 		self.assertEqual(6, len(synthDesign.InfoMessages))
 		self.assertEqual(0, len(synthDesign.WarningMessages))
 		self.assertEqual(0, len(synthDesign.CriticalWarningMessages))
 		self.assertEqual(0, len(synthDesign.ErrorMessages))
 
-		self.assertIn(RTLElaboration, synthDesign)
-		rtlElaboration = synthDesign[RTLElaboration]
+		self.assertIn(_SynthDesign.RTLElaboration, synthDesign)
+		rtlElaboration = synthDesign[_SynthDesign.RTLElaboration]
 		self.assertEqual(2, len(rtlElaboration.InfoMessages))
 		self.assertEqual(0, len(rtlElaboration.WarningMessages))
 		self.assertEqual(0, len(rtlElaboration.CriticalWarningMessages))
@@ -156,7 +154,7 @@ class SynthDesign(TestCase):
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_LoadingPart(self) -> None:
 		print()
@@ -189,15 +187,15 @@ class SynthDesign(TestCase):
 		self.assertEqual(0, len(processor.CriticalWarningMessages))
 		self.assertEqual(0, len(processor.ErrorMessages))
 
-		self.assertIn(SynthesizeDesign, processor)
-		synthDesign = processor[SynthesizeDesign]
+		self.assertIn(Synth_Design, processor)
+		synthDesign = processor[Synth_Design]
 		self.assertEqual(4, len(synthDesign.InfoMessages))
 		self.assertEqual(0, len(synthDesign.WarningMessages))
 		self.assertEqual(0, len(synthDesign.CriticalWarningMessages))
 		self.assertEqual(0, len(synthDesign.ErrorMessages))
 
-		self.assertIn(LoadingPart, synthDesign)
-		loadingPart = synthDesign[LoadingPart]
+		self.assertIn(_SynthDesign.LoadingPart, synthDesign)
+		loadingPart = synthDesign[_SynthDesign.LoadingPart]
 		self.assertEqual(0, len(loadingPart.InfoMessages))
 		self.assertEqual(0, len(loadingPart.WarningMessages))
 		self.assertEqual(0, len(loadingPart.CriticalWarningMessages))
@@ -206,7 +204,7 @@ class SynthDesign(TestCase):
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_RTLComponentStatistics(self) -> None:
 		print()
@@ -241,15 +239,15 @@ class SynthDesign(TestCase):
 		self.assertEqual(0, len(processor.CriticalWarningMessages))
 		self.assertEqual(0, len(processor.ErrorMessages))
 
-		self.assertIn(SynthesizeDesign, processor)
-		synthDesign = processor[SynthesizeDesign]
+		self.assertIn(Synth_Design, processor)
+		synthDesign = processor[Synth_Design]
 		self.assertEqual(4, len(synthDesign.InfoMessages))
 		self.assertEqual(0, len(synthDesign.WarningMessages))
 		self.assertEqual(0, len(synthDesign.CriticalWarningMessages))
 		self.assertEqual(0, len(synthDesign.ErrorMessages))
 
-		self.assertIn(RTLComponentStatistics, synthDesign)
-		rtlComponentStatistics = synthDesign[RTLComponentStatistics]
+		self.assertIn(_SynthDesign.RTLComponentStatistics, synthDesign)
+		rtlComponentStatistics = synthDesign[_SynthDesign.RTLComponentStatistics]
 		self.assertEqual(0, len(rtlComponentStatistics.InfoMessages))
 		self.assertEqual(0, len(rtlComponentStatistics.WarningMessages))
 		self.assertEqual(0, len(rtlComponentStatistics.CriticalWarningMessages))
@@ -258,7 +256,7 @@ class SynthDesign(TestCase):
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_IOInsertion(self) -> None:
 		print()
@@ -302,29 +300,29 @@ class SynthDesign(TestCase):
 		self.assertEqual(0, len(processor.CriticalWarningMessages))
 		self.assertEqual(0, len(processor.ErrorMessages))
 
-		self.assertIn(SynthesizeDesign, processor)
-		synthDesign = processor[SynthesizeDesign]
+		self.assertIn(Synth_Design, processor)
+		synthDesign = processor[Synth_Design]
 		self.assertEqual(4, len(synthDesign.InfoMessages))
 		self.assertEqual(0, len(synthDesign.WarningMessages))
 		self.assertEqual(0, len(synthDesign.CriticalWarningMessages))
 		self.assertEqual(0, len(synthDesign.ErrorMessages))
 
-		self.assertIn(IOInsertion, synthDesign)
-		ioInsertion = synthDesign[IOInsertion]
+		self.assertIn(_SynthDesign.IOInsertion, synthDesign)
+		ioInsertion = synthDesign[_SynthDesign.IOInsertion]
 		self.assertEqual(0, len(ioInsertion.InfoMessages))
 		self.assertEqual(0, len(ioInsertion.WarningMessages))
 		self.assertEqual(0, len(ioInsertion.CriticalWarningMessages))
 		self.assertEqual(0, len(ioInsertion.ErrorMessages))
 
-		self.assertIn(FlatteningBeforeIOInsertion, ioInsertion)
-		flatteningBeforeIOInsertion = synthDesign[FlatteningBeforeIOInsertion]
+		self.assertIn(_SynthDesign.FlatteningBeforeIOInsertion, ioInsertion)
+		flatteningBeforeIOInsertion = ioInsertion[_SynthDesign.FlatteningBeforeIOInsertion]
 		self.assertEqual(0, len(flatteningBeforeIOInsertion.InfoMessages))
 		self.assertEqual(0, len(flatteningBeforeIOInsertion.WarningMessages))
 		self.assertEqual(0, len(flatteningBeforeIOInsertion.CriticalWarningMessages))
 		self.assertEqual(0, len(flatteningBeforeIOInsertion.ErrorMessages))
 
-		self.assertIn(FinalNetlistCleanup, ioInsertion)
-		finalNetlistCleanup = synthDesign[FinalNetlistCleanup]
+		self.assertIn(_SynthDesign.FinalNetlistCleanup, ioInsertion)
+		finalNetlistCleanup = ioInsertion[_SynthDesign.FinalNetlistCleanup]
 		self.assertEqual(0, len(finalNetlistCleanup.InfoMessages))
 		self.assertEqual(0, len(finalNetlistCleanup.WarningMessages))
 		self.assertEqual(0, len(finalNetlistCleanup.CriticalWarningMessages))
@@ -332,7 +330,7 @@ class SynthDesign(TestCase):
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
 
 	def test_WritingSynthesisReport(self) -> None:
 		print()
@@ -424,15 +422,15 @@ class SynthDesign(TestCase):
 		self.assertEqual(0, len(processor.CriticalWarningMessages))
 		self.assertEqual(0, len(processor.ErrorMessages))
 
-		self.assertIn(SynthesizeDesign, processor)
-		synthDesign = processor[SynthesizeDesign]
+		self.assertIn(Synth_Design, processor)
+		synthDesign = processor[Synth_Design]
 		self.assertEqual(4, len(synthDesign.InfoMessages))
 		self.assertEqual(0, len(synthDesign.WarningMessages))
 		self.assertEqual(0, len(synthDesign.CriticalWarningMessages))
 		self.assertEqual(0, len(synthDesign.ErrorMessages))
 
-		self.assertIn(WritingSynthesisReport, synthDesign)
-		writingSynthesisReport = synthDesign[WritingSynthesisReport]
+		self.assertIn(_SynthDesign.WritingSynthesisReport, synthDesign)
+		writingSynthesisReport = synthDesign[_SynthDesign.WritingSynthesisReport]
 		self.assertEqual(0, len(writingSynthesisReport.InfoMessages))
 		self.assertEqual(0, len(writingSynthesisReport.WarningMessages))
 		self.assertEqual(0, len(writingSynthesisReport.CriticalWarningMessages))
@@ -444,4 +442,4 @@ class SynthDesign(TestCase):
 
 		self.assertEqual(0, len(warnings))
 		for line in processor.Lines:
-			self.assertIsInstance(line, Line)
+			self.assertIsInstance(line, VivadoLine)
